Commit con getPunto
Falta probar cosas

+++ src/java/DAO/AbstractDAO.java
@@ -74,7 +74,7 @@ public abstract class AbstractDAO {
         }
     }
 
-    protected Object find(Class clazz, Long id) {
+    protected Object find(Class clazz, Integer id) {
         Object obj = null;
         try {
             startOperation();
