+++ src/java/DAO/PuntoOmbuDAO.java
@@ -66,7 +66,7 @@ public class PuntoOmbuDAO extends AbstractDAO{
      * @param id Long correspondinete al ID del usuario
      * @return Usuario si encuentra alguno
      */
-    public PuntoOmbu find(Long id) throws DataAccessLayerException {
+    public PuntoOmbu find(Integer id) throws DataAccessLayerException {
         return (PuntoOmbu) super.find(PuntoOmbu.class, id);
     }
